airflow/requirements/requirements-airflow.txt
apache-airflow-providers-amazon
apache-airflow-providers-ssh
boto3>=1.34.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
pendulum>=3.0.0
faker>=24.0.0
